+++ scp/filesystem_test.go
@@ -23,6 +23,17 @@ func TestFilesystem(t *testing.T) {
 		requireEqualGolden(t, bts)
 	})
 
+	t.Run("copy file that does not exist to client", func(t *testing.T) {
+		is := is.New(t)
+
+		dir := t.TempDir()
+		h := NewFileSystemHandler(dir)
+
+		session := setup(t, h, nil)
+		_, err := session.CombinedOutput("scp -f a.txt")
+		is.True(err != nil)
+	})
+
 	t.Run("copy from client", func(t *testing.T) {
 		is := is.New(t)
 		dir := t.TempDir()
@@ -58,6 +69,17 @@ func TestFilesystem(t *testing.T) {
 		bts, err := session.CombinedOutput("scp -r -f a")
 		is.NoErr(err)
 		requireEqualGolden(t, bts)
+	})
+
+	t.Run("recursive copy folder that does not exist to client", func(t *testing.T) {
+		is := is.New(t)
+
+		dir := t.TempDir()
+		h := NewFileSystemHandler(dir)
+
+		session := setup(t, h, nil)
+		_, err := session.CombinedOutput("scp -r -f a")
+		is.True(err != nil)
 	})
 
 	t.Run("recursive copy from client", func(t *testing.T) {
